# openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "FastAPI",
        "version": "0.1.0"
    },
    "paths": {
...
        "/api/v1/farm/{farm_id}/seasons": {
            "get": {
                "tags": [
                    "farm"
                ],
                "summary": "Get Farm Seasons",
                "operationId": "get_farm_seasons_api_v1_farm__farm_id__seasons_get",
                "parameters": [
                    {
                        "name": "farm_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Farm Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/FarmSeason"
                                    },
                                    "title": "Response Get Farm Seasons Api V1 Farm  Farm Id  Seasons Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/season/": {
            "get": {
                "tags": [
                    "season"
                ],
                "summary": "List Seasons",
                "operationId": "list_seasons_api_v1_season__get",
                "parameters": [
                    {
                        "name": "offset",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 0,
                            "title": "Offset"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 100,
                            "default": 100,
                            "title": "Limit"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Season"
                                    },
                                    "title": "Response List Seasons Api V1 Season  Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "season"
                ],
                "summary": "Create Season",
                "operationId": "create_season_api_v1_season__post",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SeasonCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Season"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/season/{season_id}": {
            "get": {
                "tags": [
                    "season"
                ],
                "summary": "Get Season",
                "operationId": "get_season_api_v1_season__season_id__get",
                "parameters": [
                    {
                        "name": "season_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Season Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Season"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "season"
                ],
                "summary": "Update Season",
                "operationId": "update_season_api_v1_season__season_id__put",
                "parameters": [
                    {
                        "name": "season_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Season Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SeasonCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Season"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "season"
                ],
                "summary": "Delete Season",
                "operationId": "delete_season_api_v1_season__season_id__delete",
                "parameters": [
                    {
                        "name": "season_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Season Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Season"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "FarmSeason": {
                "properties": {
                    "description": {
                        "type": "string",
                        "title": "Description"
                    },
                    "year": {
                        "type": "integer",
                        "title": "Year"
                    },
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    },
                    "plantations": {
                        "items": {
                            "$ref": "#/components/schemas/Plantation"
                        },
                        "type": "array",
                        "title": "Plantations"
                    }
                },
                "type": "object",
                "required": [
                    "description",
                    "year",
                    "id",
                    "plantations"
                ],
                "title": "FarmSeason"
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
            "Plantation": {
                "properties": {
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "description": {
                        "type": "string",
                        "title": "Description"
                    },
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    }
                },
                "type": "object",
                "required": [
                    "name",
                    "description",
                    "id"
                ],
                "title": "Plantation"
            },
...
            "Season": {
                "properties": {
                    "description": {
                        "type": "string",
                        "title": "Description"
                    },
                    "year": {
                        "type": "integer",
                        "title": "Year"
                    },
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    }
                },
                "type": "object",
                "required": [
                    "description",
                    "year",
                    "id"
                ],
                "title": "Season"
            },
            "SeasonCreate": {
                "properties": {
                    "description": {
                        "type": "string",
                        "title": "Description"
                    },
                    "year": {
                        "type": "integer",
                        "title": "Year"
                    }
                },
                "type": "object",
                "required": [
                    "description",
                    "year"
                ],
                "title": "SeasonCreate"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        }
    }
}